Extract dummy imagesource
